ai/requirements.txt
langchain
openai 
python-dotenv
langchain-openai